+++ vector/src/main/java/im/vector/app/features/settings/VectorSettingsRootFragment.kt
@@ -16,7 +16,9 @@
 
 package im.vector.app.features.settings
 
+import android.os.Build
 import im.vector.app.R
+import im.vector.app.core.preference.VectorPreference
 import javax.inject.Inject
 
 class VectorSettingsRootFragment @Inject constructor() : VectorSettingsBaseFragment() {
@@ -25,6 +27,15 @@ class VectorSettingsRootFragment @Inject constructor() : VectorSettingsBaseFragm
     override val preferenceXmlRes = R.xml.vector_settings_root
 
     override fun bindPref() {
-        // Nothing to do
+        // Tint icon on API < 24
+        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.N) {
+            tintIcons()
+        }
+    }
+
+    private fun tintIcons() {
+        for (i in 0 until preferenceScreen.preferenceCount) {
+            (preferenceScreen.getPreference(i) as? VectorPreference)?.let { it.tintIcon = true }
+        }
     }
 }